requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
black==23.10.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-filter==23.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
